requirements.txt
click==6.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pip-autoremove==0.9.0
PyJWT==1.4.2
SQLAlchemy==1.1.3
Werkzeug==0.11.11
